Remove old code, tests passing
