+++ pyop3/axtree/parallel.py
@@ -1,14 +1,31 @@
 from __future__ import annotations
 
+import functools
+
 import numpy as np
 from mpi4py import MPI
 from petsc4py import PETSc
 from pyrsistent import pmap
 
 from pyop3.axtree.layout import _as_int, _axis_component_size, step_size
-from pyop3.dtypes import IntType, get_mpi_dtype
+from pyop3.dtypes import IntType, as_numpy_dtype, get_mpi_dtype
 from pyop3.extras.debug import print_with_rank
 from pyop3.utils import checked_zip, just_one, strict_int
+
+
+def reduction_op(op, invec, inoutvec, datatype):
+    dtype = as_numpy_dtype(datatype)
+    invec = np.frombuffer(invec, dtype=dtype)
+    inoutvec = np.frombuffer(inoutvec, dtype=dtype)
+    inoutvec[:] = op(invec, inoutvec)
+
+
+_contig_min_op = MPI.Op.Create(
+    functools.partial(reduction_op, np.minimum), commute=True
+)
+_contig_max_op = MPI.Op.Create(
+    functools.partial(reduction_op, np.maximum), commute=True
+)
 
 
 def partition_ghost_points(axis, sf):
